add classification strategies
